crud annuncio completato

+++ src/main/java/it/prova/myebay/service/AnnuncioServiceImpl.java
@@ -1,13 +1,18 @@
 package it.prova.myebay.service;
 
+import java.time.LocalDate;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import it.prova.myebay.exception.AnnuncioChiusoException;
 import it.prova.myebay.model.Annuncio;
+import it.prova.myebay.model.Utente;
 import it.prova.myebay.repository.annuncio.AnnuncioRepository;
+import it.prova.myebay.repository.utente.UtenteRepository;
 
 
 @Service
@@ -16,6 +21,9 @@ public class AnnuncioServiceImpl implements AnnuncioService{
 	
 	@Autowired
 	private AnnuncioRepository annuncioRepository;
+	
+	@Autowired
+	private UtenteService utenteService;
 	
 	@Override
 	@Transactional(readOnly = true)
@@ -44,12 +52,25 @@ public class AnnuncioServiceImpl implements AnnuncioService{
 	@Override
 	@Transactional
 	public void inserisciNuovo(Annuncio annuncioInstance) {
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		Utente utenteFromDb = utenteService.findByUsername(username);
+		
+		if (utenteFromDb == null)
+		throw new RuntimeException("Elemento non trovato.");
+		
+		annuncioInstance.setUtente(utenteFromDb);
+		annuncioInstance.setAperto(true);
+		annuncioInstance.setDataCreazione(LocalDate.now());
 		annuncioRepository.save(annuncioInstance);
 	}
 
 	@Override
 	@Transactional
 	public void rimuovi(Long idAnnuncio) {
+		Annuncio annuncioFromDB= this.caricaSingoloElemento(idAnnuncio);
+		if (!annuncioFromDB.isAperto()) {
+			throw new AnnuncioChiusoException();
+		}
 		annuncioRepository.deleteById(idAnnuncio);
 	}
